requirements.txt
playwright
gspread
oauth2client
requests
pillow
regex